+++ src/components/EverestUI.tsx
@@ -1,6 +1,7 @@
 import { fromJS, List } from "immutable";
 import * as moment from "moment";
 import * as React from "react";
+import * as ReactTooltip from "react-tooltip";
 
 import { CmrCollection, ICmrCollection } from "../types/CmrCollection";
 import { CmrGranule } from "../types/CmrGranule";
@@ -139,6 +140,7 @@ export class EverestUI extends React.Component<IEverestProps, IEverestState> {
     }
     return (
       <div id="everest-container">
+        <ReactTooltip effect="solid" delayShow={500} />
         <CmrDownBanner
           cmrStatusChecked={this.state.cmrStatusChecked}
           cmrStatusOk={this.state.cmrStatusOk}
@@ -154,7 +156,9 @@ export class EverestUI extends React.Component<IEverestProps, IEverestState> {
               cmrStatusOk={this.state.cmrStatusOk}
               environment={this.props.environment}
               onChange={this.handleOrderParameterChange}
-              orderParameters={this.state.orderParameters} />
+              orderParameters={this.state.orderParameters}
+              onTemporalReset={this.handleTemporalReset}
+            />
           </div>
           <div id="right-side">
             <GranuleList
@@ -337,6 +341,14 @@ export class EverestUI extends React.Component<IEverestProps, IEverestState> {
     return Promise.reject(response);
   }
 
+  private handleTemporalReset = () => {
+    const collection = this.state.orderParameters.collection;
+    this.handleOrderParameterChange({
+      temporalFilterLowerBound: collection.time_start ? moment.utc(collection.time_start) : moment.utc("20100101"),
+      temporalFilterUpperBound: collection.time_end ? moment.utc(collection.time_end) : moment.utc(),
+    });
+  }
+
   private handleOrderParameterChange = (newOrderParameters: Partial<IOrderParameters>) => {
     const orderParameters = mergeOrderParameters(this.state.orderParameters, newOrderParameters);
 
@@ -373,13 +385,7 @@ export class EverestUI extends React.Component<IEverestProps, IEverestState> {
     }
 
     const collection = cmrCollections.first();
-    const collectionSpatialCoverage = cmrBoxArrToSpatialSelection(collection.boxes);
-    this.handleOrderParameterChange({
-      collection,
-      collectionSpatialCoverage,
-      temporalFilterLowerBound: moment(collection.time_start),
-      temporalFilterUpperBound: collection.time_end ? moment(collection.time_end) : moment(),
-    });
+    this.handleCollectionChange(collection);
   }
 
   private extractOrderParametersFromLocalStorage = (): OrderParameters | null => {