+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Allow setting dates with MM/DD/YYYY format or M/D/YYYY
 * After a CMR error, display an error banner, gray out the app; when the
   user dismisses the banner restore CMR functionality
+* Add temporal reset button
 
 # v0.2.0 (2018-11-07)